gcp-vpc-move-vip.in: manage ip alias
Add a resource agent to manage ip alias in the cluster.

start:
	Check if any machine in hostlist has the alias_ip assigned and
	disassociate it.
	Assign alias_ip to the current machine.

stop:
	Disassociate the alias_ip from the current machine.

status/monitor:
	Check if alias_ip is assigned with the current machine.

---

This is a port to the following bash script to python:
https://storage.googleapis.com/sapdeploy/pacemaker-gcp/alias

The problem with the bash script is the use of gcloud whose command line
API is not stable.

ocf-tester.in results:

	> sudo ./tools/ocf-tester.in -o alias_ip='10.128.1.0/32' -o stackdriver_logging=yes -n gcp-vpc-move-vip.in heartbeat/gcp-vpc-move-vip.in
	Beginning tests for heartbeat/gcp-vpc-move-vip.in...
	./tools/ocf-tester.in: line 226: cd: @datadir@/resource-agents: No such file or directory
	close failed in file object destructor:
	sys.excepthook is missing
	lost sys.stderr
	* rc=1: Your agent produces meta-data which does not conform to ra-api-1.dtd
	Tests failed: heartbeat/gcp-vpc-move-vip.in failed 1 tests

The only test faillig is the meta-data, but all the agents that I tried
also fails on this. If this is a concern, could you please point me out
to a test which succeeds so I can check what I am doing differently?

This commit can also be viewed at:
	https://github.com/collabora-gce/resource-agents/tree/alias

Thanks